refactor(eve): move the codex transport under the extension it serves
eve/extensions/model is a staging ground for a future standalone
package, so its implementation cannot live in internal/: relocate
auth, transport, and model (plus their tests and the unsigned-jwt
test helper) to public/extensions/model/codex with relative imports.
The extension now depends only on #compiled vendored deps and two

#shared utils.
