+++ packages/eve/src/public/extensions/model/codex/auth.ts
