+++ packages/eve/src/public/extensions/model/codex/model.ts
@@ -3,10 +3,7 @@ import type {
   LanguageModelV4,
   LanguageModelV4CallOptions,
 } from "#compiled/@ai-sdk/provider/index.js";
-import {
-  createCodexFetch,
-  type CodexTransportOptions,
-} from "#internal/model-auth/endpoint/codex/transport.js";
+import { createCodexFetch, type CodexTransportOptions } from "./transport.js";
 import { isObject } from "#shared/guards.js";
 
 const CODEX_LOCAL_AUTH_API_KEY = "codex-local-auth";
